+++ main.lsf
@@ -488,7 +488,7 @@ Overlay field
     After main
     Object asm/overlay_01_021E5900.o
     Object src/field/overlay_01_021E66E4.o
-    Object src/field/overlay_01_021E6880.o
+    Object src/field/overlay_01_field_control.o
     Object asm/overlay_01_021E7FDC.o
     Object asm/overlay_01_021E8744.o
     Object asm/overlay_01_021E90C0.o
